+++ internal/signalfx-agent/pkg/monitors/prometheusexporter/prometheus.go
@@ -45,11 +45,27 @@ type Config struct {
 	// (the default).
 	MetricPath string `yaml:"metricPath" default:"/metrics"`
 
+	// Control the log level to use if a scrape failure occurs when scraping
+	// a target. Modifying this configuration is useful for less stable
+	// targets. Only the debug, info, warn, and error log levels are supported.
+	ScrapeFailureLogLevel string `yaml:"scrapeFailureLogLevel" default:"error"`
+
 	// Send all the metrics that come out of the Prometheus exporter without
 	// any filtering.  This option has no effect when using the prometheus
 	// exporter monitor directly since there is no built-in filtering, only
 	// when embedding it in other monitors.
 	SendAllMetrics bool `yaml:"sendAllMetrics"`
+}
+
+func (c *Config) Validate() error {
+	var logLevels = []string{"debug", "info", "warn", "error"}
+	for _, v := range logLevels {
+		if v == c.ScrapeFailureLogLevel {
+			return nil
+		}
+	}
+	return fmt.Errorf("invalid prometheus scrape failure log level encountered - %s - valid choices include %s",
+		c.ScrapeFailureLogLevel, strings.Join(logLevels, ", "))
 }
 
 func (c *Config) GetExtraMetrics() []string {
@@ -74,9 +90,10 @@ type Monitor struct {
 	// If true, IncludedMetrics is ignored and everything is sent.
 	SendAll bool
 
-	monitorName string
-	logger      logrus.FieldLogger
-	cancel      func()
+	monitorName        string
+	logger             logrus.FieldLogger
+	loggerFailureLevel logrus.Level
+	cancel             func()
 }
 
 type fetcher func() (io.ReadCloser, expfmt.Format, error)
@@ -84,6 +101,7 @@ type fetcher func() (io.ReadCloser, expfmt.Format, error)
 // Configure the monitor and kick off volume metric syncing
 func (m *Monitor) Configure(conf *Config) error {
 	m.logger = logrus.WithFields(logrus.Fields{"monitorType": m.monitorName, "monitorID": conf.MonitorID})
+	m.loggerFailureLevel, _ = logrus.ParseLevel(conf.ScrapeFailureLogLevel)
 
 	var bearerToken string
 
@@ -136,7 +154,8 @@ func (m *Monitor) Configure(conf *Config) error {
 	utils.RunOnInterval(ctx, func() {
 		dps, err := fetchPrometheusMetrics(fetch)
 		if err != nil {
-			m.logger.WithError(err).Error("Could not get prometheus metrics")
+			// The default log level is error, users can configure which level to use
+			m.logger.WithError(err).Log(m.loggerFailureLevel, "Could not get prometheus metrics")
 			return
 		}